[push] 멀티 전송 기능 보완

+++ moka-web/moka-web-push/src/main/java/jmnet/moka/web/push/mvc/sender/repository/PushAppRepository.java
@@ -5,11 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 /**
- * 작업 Repository
- * 2021. 2. 18.
- *
+ * 작업 Repository 2021. 2. 18.
  */
 @Repository
-public interface PushAppRepository extends JpaRepository<PushApp, Long>, PushAppRepositorySupport {
+public interface PushAppRepository extends JpaRepository<PushApp, Integer>, PushAppRepositorySupport {
 
 }
